edge cases, general fixes (#7)
* no dropout


* fix: Use dynamic tokenizer vocab size instead of hardcoded config value

- Use len(tokenizer) to get actual vocab size including special tokens
- Fix loss calculation to use model_config.vocab_size instead of cfg.model.vocab_size
- Add debug prints for tokenizer info
- Resolves padding_idx errors with tokenizers that have special tokens beyond base vocab



* style: Apply ruff formatting to validate.py



* fix: Default to bf16 instead of fp16 for mixed precision

- Change default mixed precision from fp16 to bf16 when boolean True is given
- Aligns with modern GPU capabilities (RTX 5090 and similar)
- Better numerical stability than fp16



* fix: Handle dynamic tokenizer vocab size across all training/eval phases

- Save actual vocab_size used during training in checkpoint configs
- Export uses actual vocab_size from model weights if different from config
- MTEB evaluation uses len(tokenizer) for model initialization
- Contrastive training uses len(tokenizer) for correct vocab size
- Ensures consistency between tokenizer size and model embeddings

This fixes mismatches when using tokenizers with special tokens that
extend beyond the base vocabulary (e.g., ModernBERT tokenizer).



* style: Apply ruff formatting



* feat: Add GPU-optimized vocab_size rounding for new model creation

- Add round_up_to_multiple() utility to round vocab_size to multiples of 128
- Apply rounding ONLY when creating models from scratch (random weights):
  * pretraining/trainer.py: Main pretraining from scratch
  * pretraining/trainer_phase_2.py: Phase 2 training from scratch
  * glue/train.py: Only when allow_random_weights=True (for testing)
- Do NOT apply rounding when loading existing checkpoints:
  * contrastive/trainer.py: Reverted - always loads checkpoint weights
  * evaluation/run_mteb.py: Reverted - always loads checkpoint weights
- Improves GPU efficiency by ensuring embedding matrices have optimal dimensions
- Example: ModernBERT tokenizer (50373 tokens) -> rounded to 50432



* style: Apply ruff formatting



* fix: Update config vocab_size early to propagate to all outputs

- Update cfg.model.vocab_size and cfg.tokenizer.vocab_size immediately after rounding
- This ensures the rounded value is used in:
  * Checkpoint config.yaml files
  * WandB config logging
  * Model state dict
  * Any other outputs that reference the config
- Remove duplicate update that happened too late in checkpoint saving
- Apply same pattern to trainer_phase_2.py

The rounded vocab_size (e.g., 50432 for ModernBERT) is now the single source
of truth throughout the training pipeline.



* style: Apply ruff formatting



* docs: Add vocabulary size optimization documentation

- Document automatic vocab_size rounding to multiples of 128
- Explain the flow from config to runtime to exports
- Provide ModernBERT tokenizer as concrete example
- Clarify this only applies when training from scratch



* fix: Set dropout_prob to 0.0 in all pretraining configs

Modern encoder-only models no longer use dropout during pretraining,
following the trend set by recent models like ModernBERT and others.
This improves training stability and model performance.



* fix: Set dropout_prob to 0.0 in all pretraining test configs

Updated all test configs in tests/configs/pretraining/ to use dropout_prob: 0.0
to match the main pretraining configs. This ensures tests reflect actual
training behavior where dropout is no longer used during pretraining.



* feat: Add logging warnings when vocab_size is rounded for GPU efficiency

- Add explicit logging.warning when vocab_size needs rounding to multiple of 128
- Import logging module in trainer_phase_2.py
- Helps users understand why their tokenizer vocab_size differs from model vocab_size

* refactor: Move vocab_size resolution to config preprocessing

Fix WandB logging incorrect vocab_size by resolving dynamic values
during config loading instead of in trainer. This ensures all
downstream consumers (WandB, model, etc.) get consistent config.

Changes:
- Add ConfigLoader.preprocess_config() for resolving dynamic values
- Move vocab_size rounding from trainer.py to config preprocessing
- Update load_config_from_args() to apply preprocessing by default
- Remove redundant vocab_size logic from trainer



* 🐛 fix mlm example generation


---------

+++ src/neobert/validation/validators.py
@@ -94,7 +94,7 @@ def validate_glue_config(cfg: Any) -> None:
             # Handle both string and boolean values
             mixed_precision = cfg.trainer.mixed_precision
             if isinstance(mixed_precision, bool):
-                mixed_precision = "no" if not mixed_precision else "fp16"
+                mixed_precision = "no" if not mixed_precision else "bf16"
             if mixed_precision not in valid_precision:
                 errors.append(
                     f"Invalid mixed_precision: {cfg.trainer.mixed_precision}. "
